+++ dojo/test/ui/views.py
@@ -25,8 +25,8 @@ from django.views.decorators.vary import vary_on_cookie
 import dojo.finding.helper as finding_helper
 from dojo.authorization.authorization import user_has_permission_or_403
 from dojo.engagement.queries import get_authorized_engagements
-from dojo.filters import FindingFilter, FindingFilterWithoutObjectLookups, TemplateFindingFilter
 from dojo.finding.queries import prefetch_for_findings
+from dojo.finding.ui.filters import FindingFilter, FindingFilterWithoutObjectLookups, TemplateFindingFilter
 from dojo.finding.views import find_available_notetypes
 from dojo.forms import (
     AddFindingForm,
