+++ unittests/test_filter_finding_mitigation.py
@@ -3,7 +3,8 @@ import datetime
 from django.test import TestCase
 from django.utils import timezone
 
-from dojo.filters import ApiFindingFilter, FindingFilterHelper
+from dojo.filters import ApiFindingFilter
+from dojo.finding.ui.filters import FindingFilterHelper
 from dojo.models import (
     Dojo_User,
     Engagement,
